requirements.txt
streamlit>=1.31.0
openai>=1.12.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken==0.5.1
langchain==0.0.300 